requirements.txt
asgiref==3.7.2
Django==3.2.25
django-filter==23.5
djangorestframework==3.14.0
Faker==18.13.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
validate-docbr==1.10.0
